+++ lib/thor-tropo/uploader.rb
@@ -13,6 +13,7 @@ module ThorTropo
         @debug_mode   = opts[:debug_mode]
         @connection   = setup_connection(opts[:access_key],opts[:secret_key])
         @bucket_name  = get_bucket( opts[:bucket] || 'artifacts.voxeolabs.net' )
+        @use_iam      = opts[:use_iam]
       end
 
       def upload(opts={})
@@ -56,11 +57,19 @@ module ThorTropo
       private
 
       def setup_connection(u,p)
-        Fog::Storage.new({
+        if @use_iam
+          opts = {
+            :provider        => 'AWS',
+            :use_iam_profile => true
+          }
+        else
+          opts = {
             :provider                 => 'AWS',
             :aws_access_key_id        => u,
             :aws_secret_access_key    => p
-        })
+          }
+        end
+        Fog::Storage.new(opts)
       end
 
       def get_bucket(bucket_name)