+++ changelog.md
@@ -1,3 +1,13 @@
+# 2.0.0
+
+* Support multiple cookbooks via .deployer
+* Support IAM profiles
+* Support --version / -v 
+* Support --help / -h
+* Better logging
+* Support git tagging via seperate task
+* Support cleaning berksfile cache
+
 # 1.4.0
 
 * Moved cookbook finder into it's own method, SRP :)